+++ react/components/Layout/index.js
@@ -3,13 +3,19 @@ import PropTypes from 'prop-types'
 
 class Layout extends Component {
   render() {
-    const { fullWidth, pageHeader, children } = this.props
+    const { fullWidth, pageHeader, stretch, children } = this.props
 
     return (
       <div className="styleguide__layout flex justify-center pb7 bg-muted-5 min-h-100">
-        <div className={fullWidth ? 'w-100' : 'w-100 mw8'}>
+        <div
+          className={`h-100 w-100 flex flex-column ${!fullWidth ? 'mw8' : ''}`}>
           {pageHeader}
-          <div className="layout__container ph7-ns">{children}</div>
+          <div
+            className={`layout__container ph7-ns ${
+              stretch ? 'flex flex-column flex-auto' : ''
+            }`}>
+            {children}
+          </div>
         </div>
       </div>
     )
@@ -18,6 +24,7 @@ class Layout extends Component {
 
 Layout.defaultProps = {
   fullWidth: false,
+  stretch: false,
 }
 
 Layout.propTypes = {
@@ -27,6 +34,8 @@ Layout.propTypes = {
   fullWidth: PropTypes.bool,
   /** A PageHeader component slot */
   pageHeader: PropTypes.node.isRequired,
+  /** If the content should stretch and fill the container height */
+  stretch: PropTypes.bool,
 }
 
 export default Layout